For github/requirements.txt
altair==4.2.2
joblib==1.2.0
keras==2.12.0
matplotlib==3.6.2
numpy==1.23.1
openai==0.27.0
pandas==1.5.1
Pillow==9.5.0
plotly==5.14.0
streamlit==1.19.0
tensorflow==2.12.0
statsmodels 
scikit-learn	
inotify	
